+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.18.0-RELEASE
+
+* Add support for new security features when sending a file by email:
+  * `confirmEmailBeforeDownload` can be set to `True` to require the user to enter their email address before accessing the file.
+  * `retentionPeriod` can be set to `<1-78> weeks` to set how long the file should be made available.
+
+
 ## 3.17.3-RELEASE
 * Removed unused commons-cli dependency